use termion for all non-windows platforms
