+++ Cargo.toml
@@ -11,11 +11,13 @@ license = "MIT"
 [dependencies]
 libc = "0.2.9"
 time = "0.1.35"
+
+[target.'cfg(target_os = "windows")'.dependencies]
 winapi = "0.2"
 kernel32-sys = "0.2"
 
-[target.'cfg(target_os = "redox")'.dependencies]
-termion = "1.4"
+[target.'cfg(not(target_os = "windows"))'.dependencies]
+termion = "1.5"
 
 [dev-dependencies]
 rand = "0.3.14"
